Start Round Button refactoring; Conditional rendering on rounds history
